PC側の実装の削除
